+++ native/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the native code for "zowex" are documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## Recent Changes
+
+- `c`: Flipped to and from encoding arguments in error message to be accurate. [#1106](https://github.com/zowe/zowex/pull/1106)
+
 ## `0.8.0`
 
 - `c`: Stripped ASA control characters when reading spool files. [#114](https://github.com/zowe/zowex/issues/114)